Share common validations between all advocate claim types
With the introduction of AGFS supplementary claims
there is a lot of shared validation logic between
the different types.

Rather than use an intermediate class for
advocate claim validators this uses a two mixins
that can be added to the different types as needed.

Advocate (final) claims and supplementary claims
share alot of logic related to defendant uplifts
and all advocate claim types share other mandatory
type validation.

Note: this has added an extra validation to advocate
interim claims (validate_creator), which, to my mind
was missing from it in any event.
