+++ app/validators/claim/advocate_claim_validator.rb
@@ -1,4 +1,7 @@
 class Claim::AdvocateClaimValidator < Claim::BaseClaimValidator
+  include Claim::AdvocateClaimCommonValidations
+  include Claim::DefendantUpliftValidations
+
   def self.fields_for_steps
     {
       case_details: %i[
@@ -40,101 +43,9 @@ class Claim::AdvocateClaimValidator < Claim::BaseClaimValidator
 
   private
 
-  def supplier_number_regex
-    ExternalUser::SUPPLIER_NUMBER_REGEX
-  end
-
-  def validate_creator
-    super if defined?(super)
-    validate_has_role(@record.creator.try(:provider),
-                      :agfs,
-                      :creator,
-                      'must be from a provider with permission to submit AGFS claims')
-  end
-
-  def validate_advocate_category
-    validate_presence(:advocate_category, 'blank')
-    return if @record.advocate_category.blank?
-    validate_inclusion(:advocate_category, @record.eligible_advocate_categories, I18n.t('validators.advocate.category'))
-  end
-
   def validate_offence
     return if fixed_fee_case?
     error_message = @record.agfs_reform? ? 'new_blank' : 'blank'
     validate_presence(:offence, error_message)
-  end
-
-  def validate_case_concluded_at
-    validate_absence(:case_concluded_at, 'present')
-  end
-
-  def validate_supplier_number
-    validate_pattern(:supplier_number, supplier_number_regex, 'invalid')
-  end
-
-  def validate_defendant_uplifts_basic_fees
-    return if @record.from_api? || @record.fixed_fee_case?
-    return unless defendant_uplifts_greater_than?(defendant_uplifts_basic_fees_counts, number_of_defendants)
-    add_error(:base, 'defendant_uplifts_basic_fees_mismatch')
-  end
-
-  def validate_defendant_uplifts_fixed_fees
-    return if @record.from_api? || !@record.fixed_fee_case?
-    return unless defendant_uplifts_greater_than?(defendant_uplifts_fixed_fees_counts, number_of_defendants)
-    position = @record.fixed_fees.find_index(&:defendant_uplift?) + 1
-    add_error("fixed_fee_#{position}_quantity", 'defendant_uplifts_fixed_fees_mismatch')
-  end
-
-  def validate_defendant_uplifts_misc_fees
-    return if @record.from_api?
-    return unless defendant_uplifts_greater_than?(defendant_uplifts_misc_fees_counts, number_of_defendants)
-    add_error(:base, 'defendant_uplifts_misc_fees_mismatch')
-  end
-
-  def number_of_defendants
-    @record.defendants.reject(&:marked_for_destruction?).size
-  end
-
-  # we add one because uplift quantities reflect the number of "additional" defendants
-  def defendant_uplifts_greater_than?(defendant_uplifts_counts, no_of_defendants)
-    defendant_uplifts_counts.map(&:to_i).any? { |sum| sum + 1 > no_of_defendants }
-  end
-
-  def defendant_uplifts_basic_fees
-    defendant_uplifts_fees(@record.basic_fees)
-  end
-
-  def defendant_uplifts_fixed_fees
-    defendant_uplifts_fees(@record.fixed_fees)
-  end
-
-  def defendant_uplifts_misc_fees
-    defendant_uplifts_fees(@record.misc_fees)
-  end
-
-  def defendant_uplifts_fees(fees)
-    fees.select do |fee|
-      !fee.marked_for_destruction? &&
-        fee&.defendant_uplift?
-    end
-  end
-
-  def defendant_uplifts_basic_fees_counts
-    defendant_uplifts_counts_for(defendant_uplifts_basic_fees)
-  end
-
-  def defendant_uplifts_fixed_fees_counts
-    defendant_uplifts_counts_for(defendant_uplifts_fixed_fees)
-  end
-
-  def defendant_uplifts_misc_fees_counts
-    defendant_uplifts_counts_for(defendant_uplifts_misc_fees)
-  end
-
-  def defendant_uplifts_counts_for(defendant_uplifts)
-    defendant_uplifts.each_with_object({}) do |fee, res|
-      res[fee.fee_type.unique_code] ||= []
-      res[fee.fee_type.unique_code] << fee.quantity
-    end.values.map(&:sum)
   end
 end
